HW09 исправленный + AccountService + ATM

+++ HW09-jdbc-template/src/main/java/ru/otus/hw09/dao/Account.java
@@ -9,7 +9,7 @@ public class Account {
     private final String type;
     private float rest;
 
-    public Account(long no, String type, int rest) {
+    public Account(long no, String type, float rest) {
         this.no = no;
         this.type = type;
         this.rest = rest;
